AuthProvider & useAuth
Refactored authentication system by separating context provider and custom hook (AuthProvider & useAuth) to improve code organization and scalability.
